mcq generator/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
pdfplumber
python-docx
google-genai
python-dotenv
